+++ server/src-lib/Hasura/Server/Telemetry.hs
@@ -196,7 +196,7 @@ computeActionsMetrics actionCache =
 
         typeRelationships =
           length . L.nub . concatMap
-          (map _trName . maybe [] toList . _otdRelationships . _aotDefinition . _aiOutputObject) $
+          (map _trName . maybe [] toList . _otdRelationships . _aotDefinition . snd . _aiOutputObject) $
           actions
 
 -- | Logging related